init commit, experimenting with Ace code editor
This initial commit includes one component that uses the Ace code editor library (used on c9.io) to create a simple code editor.
As an experiment I was able to add a font size changer that adjusts the code editor's font size.
Clicking the run button simply logs the editor object to the console for inspection.